Add files via upload
The data set shows the general trends of usage of the osXtern application at various time intervals. This will be important since it provides information about the peak periods of usage as well as providing the time at which the usage of the osXtern application is low. The data also shows a directly proportional relationship between the number of users and the x- coordinate values. It also shows the relationship between the y - coordinate and the number of users. In both scenarios, they is a directly proportional relationship shown. This will also be helpful for the developers of the osXtern application since it will provide them with the various dependence and interdependence trends that exist on the data sets. The relationship between the x and y coordinate is directly proportional. This is important since a highly positive correlation plot can be plotted using these x and y values. A lot of information can then be ascertained from the scatter plot by inserting a line of best fit on the scatter plot and deducing the linear equation.